requirements.txt
fastapi
uvicorn
tensorflow
opencv-python-headless
numpy
stockfish
pillow
python-multipart
